+++ CotEditor/Sources/Document.swift
@@ -838,7 +838,7 @@ final class Document: NSDocument, AdditionalDocumentPreparing, EncodingChanging 
         } catch {
             self.readingEncoding = nil
             
-            throw ReinterpretationError.reinterpretationFailed(fileURL: fileURL, encoding: encoding)
+            throw ReinterpretationError.reinterpretationFailed(encoding)
         }
     }
     
@@ -1029,7 +1029,7 @@ final class Document: NSDocument, AdditionalDocumentPreparing, EncodingChanging 
     
     /// Changes the text encoding by asking options to the user.
     ///
-    /// - Parameter fileEncoding: The file encoding to change.
+    /// - Parameter fileEncoding: The text encoding to change.
     @MainActor private func askChangingEncoding(to fileEncoding: FileEncoding) {
         
         assert(Thread.isMainThread)
@@ -1053,7 +1053,7 @@ final class Document: NSDocument, AdditionalDocumentPreparing, EncodingChanging 
             
             // ask whether just change the encoding or reinterpret document file
             let alert = NSAlert()
-            alert.messageText = String(localized: "Text encoding")
+            alert.messageText = String(localized: "Text encoding change")
             alert.informativeText = String(localized: "Do you want to convert or reinterpret this document using “\(fileEncoding.localizedName)”?")
             alert.addButton(withTitle: String(localized: "Convert"))
             alert.addButton(withTitle: String(localized: "Reinterpret"))
@@ -1083,7 +1083,7 @@ final class Document: NSDocument, AdditionalDocumentPreparing, EncodingChanging 
                         if self.isDocumentEdited {
                             let alert = NSAlert()
                             alert.messageText = String(localized: "The document has unsaved changes.")
-                            alert.informativeText = String(localized: "Do you want to discard the changes and reopen the document using “\(fileEncoding.localizedName)”?",
+                            alert.informativeText = String(localized: "Are you sure you want to discard the changes and reopen the document using “\(fileEncoding.localizedName)”?",
                                                            comment: "%@ is an encoding name")
                             alert.addButton(withTitle: String(localized: "Cancel"))
                             alert.addButton(withTitle: String(localized: "Discard Changes"))
@@ -1237,7 +1237,7 @@ final class Document: NSDocument, AdditionalDocumentPreparing, EncodingChanging 
 private enum ReinterpretationError: LocalizedError {
     
     case noFile
-    case reinterpretationFailed(fileURL: URL, encoding: String.Encoding)
+    case reinterpretationFailed(String.Encoding)
     
     
     var errorDescription: String? {
@@ -1246,8 +1246,8 @@ private enum ReinterpretationError: LocalizedError {
             case .noFile:
                 String(localized: "The document doesn’t have a file to reinterpret.")
                 
-            case .reinterpretationFailed(let fileURL, let encoding):
-                String(localized: "The file “\(fileURL.lastPathComponent)” couldn’t be reinterpreted using text encoding “\(String.localizedName(of: encoding)).”")
+            case .reinterpretationFailed(let encoding):
+                String(localized: "The document could not be reinterpreted using text encoding “\(String.localizedName(of: encoding)).”")
         }
     }
     
@@ -1259,7 +1259,7 @@ private enum ReinterpretationError: LocalizedError {
                 nil
                 
             case .reinterpretationFailed:
-                String(localized: "The file may have been saved using a different text encoding, or it may not be a text file.")
+                String(localized: "The document may have been saved using a different text encoding, or it may not be a text file.")
         }
     }
 }
@@ -1273,13 +1273,13 @@ struct LossyEncodingError: LocalizedError, RecoverableError {
     
     var errorDescription: String? {
         
-        String(localized: "Some characters would have to be changed or deleted in saving as “\(self.encoding.localizedName).”")
+        String(localized: "The document contains characters incompatible with “\(self.encoding.localizedName).”")
     }
     
     
     var recoverySuggestion: String? {
         
-        String(localized: "Do you want to change encoding and show incompatible characters?")
+        String(localized: "Incompatible characters are substituted or deleted in saving. Do you want to change the text encoding and review the incompatible characters?")
     }
     
     
@@ -1321,14 +1321,14 @@ private struct SavingError: LocalizedError, RecoverableError {
         
         switch self.code {
             case .lossyEncoding(let fileEncoding):
-                String(localized: "Some characters would have to be changed or deleted in saving as “\(fileEncoding.localizedName).”")
+                String(localized: "The document contains characters incompatible with “\(fileEncoding.localizedName).”")
         }
     }
     
     
     var recoverySuggestion: String? {
         
-        String(localized: "Do you want to continue processing?")
+        String(localized: "Incompatible characters are substituted or deleted in saving. Do you want to continue processing?")
     }
     
     
@@ -1337,7 +1337,7 @@ private struct SavingError: LocalizedError, RecoverableError {
         switch self.code {
             case .lossyEncoding:
                 [String(localized: "Save Available Text"),
-                 String(localized: "Show Incompatible Characters"),
+                 String(localized: "Review Incompatible Characters"),
                  String(localized: "Cancel")]
         }
     }